Quiz 도메인에 맞게 분리
sendQuiz() : 문제 출제 함수
checkQuiz() : 정답 확인 함수
- 문제 출제와 정답 확인 도메인은 Quiz가 담당하게 수정
